backend/requirements.txt
flask
flask-cors
joblib
numpy
scikit-learn
pytest
pytest-cov
pytest-mock
requests-mock
bandit
safety
flake8
black
isort
mypy
locust
supabase
python-dotenv
